Правки по замечаниям из пулреквеста
